Move parameters to shared module.
