--- plotter/plotrequest.go
@@ -1,10 +1,0 @@
-package main
-
-import "github.com/ebeeton/buddhabrot-go/plotter/parameters"
-
-// Type PlotRequest represents a request for a Buddhabrot image. It contains the
-// database ID for the result, and the parameters used to perform the plot.
-type PlotRequest struct {
-	Id   int64
-	Plot parameters.Plot
-}
