+++ shared/parameters/plotrequest.go
@@ -1,10 +1,9 @@
-package main
-
-import "github.com/ebeeton/buddhabrot-go/parameters"
+// Package parameters defines parameters used to plot Buddhabrot images.
+package parameters
 
 // Type PlotRequest represents a request for a Buddhabrot image. It contains the
 // database ID for the result, and the parameters used to perform the plot.
 type PlotRequest struct {
 	Id   int64
-	Plot parameters.Plot
+	Plot Plot
 }
