+++ shared/parameters/stop.go
@@ -5,6 +5,6 @@ package parameters
 // Stop is a hexadecimal color value and its position in a linear gradient from
 // 0 to 1.
 type Stop struct {
-	Color    string
-	Position float64
+	Color    string  `validate:"hexcolor"`
+	Position float64 `validate:"gte=0,lte=1"`
 }
